Create Docker image for Crucible CI precheck stage
I created a precheck stage in Crucible CI that checks for formatting
issues in python code. As a result, the Black formatting tool was added
as a dependency to the CI pipeline. After a discussion with Adam S., we
determined that the best method to satisfy this dependency was to
create a Docker container for the precheck stage.

Therefore, I added the Dockerfile and requirements.txt.